refactor(#701): split src/daemon/ci_watch.rs (4829 LOC) into submodules
Pure structural split — zero behavior change. Original file decomposed into
six co-located files under src/daemon/ci_watch/:

- mod.rs       — public API re-exports + WATCH_TTL_HOURS
- registry.rs  — ci_watches_dir, parse_subscribers, watch_filename,
                 remove_watch, update_watch_state* (atomic helpers)
- provider.rs  — CiAuth, CiHttpClient, CiRun, CiPollResult, PrState,
                 CiProvider trait + GitHub/GitLab/Bitbucket impls,
                 github_token_warning*, detect_provider_from_remote
- sweep.rs     — STALL_THRESHOLD, stall fan-out, gc_stale_watches,
                 startup_sweep
- poller.rs    — adaptive_interval, watch_is_due, dedup/aggregate helpers,
                 ci_check_repo, check_ci_watches_with_provider, shared_ci_runtime,
                 #[cfg(test)] mod tests (107 tests, moved verbatim)
- watcher.rs   — check_ci_watches entry point + provider factory glue

Visibility adjustments (minimum to satisfy new module boundaries):
- pub(super) on CiHttpClient.base_url + provider.http fields so the
  cross-module test mod in poller can keep its existing constructor
  URL assertions
- pub(super) on registry::update_watch_state* and sweep::bump_/clear_
  helpers — sibling-only callers
- pub(crate) on registry::parse_subscribers, sweep::STALL_THRESHOLD
  preserved verbatim

Test-side changes:
- moved #[cfg(test)] mod tests block as-is into poller.rs (where
  ci_check_repo lives) per dispatch contract; preamble adds explicit
  cross-module use statements for sibling-only items
- include_str! fixture paths bumped one ../ level (../../tests/ → ../../../tests/)
- stale_pr_not_classified_as_terminal: include_str!("ci_watch.rs") →
  include_str!("provider.rs") — the closed_at freshness check now lives
  in GitHubCiProvider::check_pr_terminal (provider.rs)
- tests/spawn_rationale_audit.rs: daemon/ci_watch.rs → daemon/ci_watch/poller.rs
  (shared_ci_runtime().spawn site moved with check_ci_watches_with_provider)

External callers (mcp/handlers/ci, worktree_pool, app/mod, etc.) unchanged:
crate::daemon::ci_watch::{check_ci_watches, ci_watches_dir,
github_token_warning_from_env, parse_subscribers, startup_sweep,
watch_filename, WATCH_TTL_HOURS} all preserved verbatim via mod.rs
pub use re-exports.

Verification:
- cargo build: clean (no warnings)
- cargo build --tests: clean
- cargo clippy --bin agend-terminal --tests: clean
- cargo test --bin agend-terminal 'daemon::ci_watch::': 107 passed; 0 failed
- cargo test (full suite): 2076 passed lib + integration tests green
- file_size_invariant + spawn_rationale_audit: green after path updates

scope follows dispatch spec (#701 BLOCK 1 / Group B slot 1).


Agend-Agent: dev-arch-1
Agend-Task: t-20260513161023246190-2
Agend-Branch: refactor/701-split-ci-watch
Agend-Issued-At: 2026-05-13T16:11:30.602226+00:00
